refactor(ecstore): route replication errors
