+++ crates/ecstore/src/bucket/replication/mod.rs
@@ -16,6 +16,7 @@ mod config;
 pub mod datatypes;
 mod replication_bandwidth_boundary;
 mod replication_config_store;
+mod replication_error_boundary;
 mod replication_event_sink;
 mod replication_filemeta_boundary;
 mod replication_lock_boundary;
