+++ crates/ecstore/src/bucket/replication/replication_storage_boundary.rs
@@ -12,10 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::error::Error;
 use rustfs_filemeta::FileInfo;
 use tokio_util::sync::CancellationToken;
 
+use super::replication_error_boundary::Error;
 pub(crate) use crate::client::api_get_options::{AdvancedGetOptions, StatObjectOptions};
 pub(crate) use crate::object_api::{GetObjectReader, ObjectInfo, ObjectOptions, PutObjReader};
 pub(crate) use crate::storage_api_contracts::list::{
